Fixed responsive design | Now it works on all screens.
The last commit has a huge mistake, and is that i had a 50% zoom on the browser, so I design all for an screen with a 50% zoom. I realized this mistake and update the fonts size and heights for the grid to be responsible. Updated files BentoGrid.jsx Containers.jsx